fix(sitemanage): Xlint batch 1 — serialVersionUID + real generics (#2032)
Clear first coherent cluster of main-source project -Xlint diagnostics in
sitemanage (~328→259 under uncapped -Xmaxwarns):

- serialVersionUID on JAXB/JSON ArrayList wrappers and simple exceptions
- IPSExtensionService/PSExtensionService Iterator generics
- PSDispatchingPathService List<PSItemProperties> path APIs
- PSSitePublishServiceHelper List<Integer> + typed native-query helper
- PSItemService SecureKeyRotationListener Set<Integer>
- Remove redundant casts / Collections.EMPTY_LIST

Tests: PSSerializableListWrappersTest, PSCategoryControlUtilsConvertTest,
PSExtensionServiceTypingTest.

Partial for #2032. Refs #2200.

> Co-Authored by Grok Build using grok-4.5 with agent main.

+++ projects/sitemanage/src/main/java/com/percussion/assetmanagement/service/impl/PSAssetService.java
@@ -1755,7 +1755,7 @@ public class PSAssetService extends PSAbstractFullDataService<PSAsset, PSAssetSu
     IPSCmsObjectMgr objMgr = PSCmsObjectMgrLocator.getObjectManager();
     try {
       objMgr.changeWorkflowForItem(
-          (int) idMapper.getGuid(assetId).getUUID(), wfId, steppedWfMetadata.getSystemStatesList());
+          idMapper.getGuid(assetId).getUUID(), wfId, steppedWfMetadata.getSystemStatesList());
     } catch (PSORMException e) {
       String msg =
           "Failed to assign workflow for asset with ID \""
